Update README.md
Working and Features
In this website, we will get the live data of corona affected humans in India. The data is displayed using the API. This site shows the number of confirmed cases, number of recovered cases, and number of death cases in the state as well as in the cities. It also shows the containment zones. There is another slide of a named blog in which news and important instructions related to COVID-19 will be displayed. The other feature is the test, in which there are a small question and answer round through which you would roughly know about the chance of virus in you.